cloudfunctions/animalPredict/requirements.txt
tensorflow-cpu==2.10.0
pillow==9.5.0
numpy==1.21.6
opencv-python-headless==4.5.5.64